Update to stateTrans functions
baStory.R: Stand ID is printed in debug.
canSizCls.R: Reconfigured calculation of CC for size classes. Added logic for adjusting cansizcl when type is 2 or 3. Stand ID is printed in debug. Round CC function was included in file but is not currently being used.
dbInput.R: Added allRuns as an argument and adjusted logic to account for this new variable.
domType.R:  Stand ID is printed in debug.
main.R: allRuns is now an argument. Couple of checks on runTitles argument were removed. Additional logic was included to support allRuns when TRUE.