Update to vegClass
baStory: changed input arguments and a portion of logic for baStory function so that it can used outside of main function for testing.

canSizCls: changed input arguments and a portion of logic for canSizCls function so that it can used outside of main function for testing.

commons.R: File with suite of vectors used in main.R. This file replaces the supportSP.Rdata file that was previously used.

dataTools.R: modified return statement in dbCombine function.
dbInput.R: minor reformatting of cat statements.

domType: changed input arguments and a portion of logic for domType function so that it can used outside of main function for testing.

main.R: removed use of supportSP.Rdata. Changed calls to canSizCls, baStory, and domType. Removed standAttr vector.

standAttr.R: New file that contains functions for calculating CC, TPA, and BA of plot/stand.